Migrate KebabOption
