+++ frontend/packages/console-dynamic-plugin-sdk/src/extensions/console-types.ts
@@ -585,3 +585,21 @@ export type SelfSubjectAccessReviewKind = {
     evaluationError?: string;
   };
 };
+
+export type KebabOption = {
+  hidden?: boolean;
+  label?: React.ReactNode;
+  labelKey?: string;
+  labelKind?: { [key: string]: string | string[] };
+  href?: string;
+  callback?: () => any;
+  accessReview?: AccessReviewResourceAttributes;
+  isDisabled?: boolean;
+  tooltip?: string;
+  tooltipKey?: string;
+  // a `/` separated string where each segment denotes a new sub menu entry
+  // Eg. `Menu 1/Menu 2/Menu 3`
+  path?: string;
+  pathKey?: string;
+  icon?: React.ReactNode;
+};
